feat(cli): command sign (#103)
**Motivation**

Expand CLI commands: implementation of the sign command

**Description**

Implementation of the command sign for rex that given a message in bytes
that will be passed as a hex string (with or without the 0x prefix) and
a private key (also a hex string), prints the signed message (a.k.a.
signature).

The signature is computed as a 0x45 signature, as described in
[EIP-191](https://eips.ethereum.org/EIPS/eip-191), and has an extra byte
concatenated at the end that's a scalar value added to the signatures
parity, as described in the [Yellow
Paper](https://ethereum.github.io/yellowpaper/paper.pdf) Section 4.2 in
the specification of a transaction's w field.

To try the command you should use it like this:

```
rex sign <MESSAGE> <PRIVATE_KEY>
```

For example:

```
rex sign 0x0101010101 0x052d7e212b30b505f337b0be5cf953c3d3175724f337b0be5cf953c3d3175724
```

Closes #24

---------

Signed-off-by: cdiielsi <49721261+cdiielsi@users.noreply.github.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>