feat(sdk): implement calldata decoding (#134)
**Motivation**

It's useful to be able to decode calldata for debugging. This
functionality is also useful for the ethrex CLI.

**Description**

Implements calldata decoding and makes use of relevant existing tests. A
decode-calldata command is impemented for ease of testing.

For symmetry, a a encode-calldata command is also implemented since the
logic already exists.

---------

Signed-off-by: Lucas Fiegl <iovoid@users.noreply.github.com>