fix(sdk): fix noisy deploy (#149)
**Motivation**

The deploy command has a `--print-address` argument, that ensures only
the address of the deployed contract is printed.

In #148, a `wait_for_transaction_receipt` call was added to the sdk
deploy path, which is noisy and redundant (deploy calls an optionally
silent version of wait_for_transaction_receipt).

**Description**

Removes the extra `wait_for_transaction_receipt` call.