feat(cli): add cli integration test to ci (#174)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
A ci job for the rex cli was needed.

**Description**

This PR includes a new job in the Integration Test CI which tests the
cli commands of rex.

This test follows the same logic as the SDK's integration test, but
using command line arguments.

To achieve running the CLI from a rust testfile, the crate
`std::process::Command` was used.


## Test structure

First, a deposit is performed from a rich L1 wallet and the amounts are
checked after it.

Then, a transfer from this wallet to another wallet is done and all the
amounts are verified.

And lastly, 5 withdrawals are executed and the funds are back on the L1.

---------

Signed-off-by: Mechardo <30327624+Mechanix97@users.noreply.github.com>